Add drag-and-drop attachment selector to editor plugin
Introduces new React components (AttachmentSelector, AttachmentItem, DropZone) for selecting and reordering image attachments in the editor plugin. Adds associated stylesheets for both LTR and RTL layouts. Updates PHP to enqueue the new styles. The UI now allows users to visually select, order, and manage up to four attachments for federation, improving usability and clarity.
